test: replace pytest e2e with BDD behave feature specs
Switch e2e tests from pytest to behave with Gherkin Given/When/Then
syntax. Feature file at tests/e2e/features/knowledge_base.feature
covers 7 scenarios, 48 steps:

- Empty knowledge base reports zero chunks
- Index documents from configured directories
- List indexed documents
- Search the knowledge base
- Retrieve a specific document
- Server status reflects indexed data
- Show resolved configuration

Run with: uv run behave tests/e2e/features/
