+++ pyproject.toml
@@ -17,6 +17,7 @@ markers = [
 
 [dependency-groups]
 dev = [
+    "behave>=1.3.3",
     "pytest>=8.0",
     "pytest-asyncio>=0.24",
 ]
